package/systemd: fix build with glibc 2.39
When glibc was bumped to version 2.39 in commit
b5680f53d60acf8ff6010082f873438a39bd5d97 it removed the deprecated
libcrypt support.

As glibc's libcrypt was providing systemd's libcrypt dependency this
broke any systemd build using glibc version 2.39.

To fix this add the libxcrypt dependency to systemd which is the
preferred way of providing libcrypt support in systemd as the glibc
variant is only used as a fallback due to being deprecated.

We should also have host-systemd depend on host-libxcrypt in case
the host system does not provide glibc with libcrypt support.

Fixes:
 - http://autobuild.buildroot.net/results/36e/36efcfc76c015c9b4c955c13afa0f81a98b529f4

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>

Change-Id: I178cfcda133a1083bc010d2a0ee1623d4d62b7ea
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>