utils/brmake: simplify redirection
We want to redirect both stdout and stderr to the log, so just do both
redirections at once. This simplifies the call to make, which was
inherited from afdb545b2847 (tools: new tool to filter the output of
make) when brmake was using unbuffer (and even then it was not entirely
waranted to do the redirection that way).

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>
(cherry picked from commit d21aeb44e06eb8259179edb3f7e330a07f8b908f)

Change-Id: I378446931fcc18b4e6ebe80b91344877deaf51a0
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>